add workers to influx and redis writing

+++ apps/telemetry/grpc-parser/internal/parser/milesight/parser.go
@@ -119,7 +119,7 @@ func ParseMilesightUplink(payload []byte, model string) (*util.ParsedData, error
 					_ = pulseConv
 					// IEEE-754 float32 pulse counter (little-endian)
 					raw := binary.LittleEndian.Uint32(payload[i+4 : i+8])
-					pulseCount := float64(math.Float32frombits(raw))
+					pulseCount := math.Float32frombits(raw)
 					// dp.Fields["pulse_count"] = util.FieldsValue{Key: "pulse_count", Value: pulseCount}
 					dp.Fields["pulse_count"] = pulseCount
